+++ backlib/__init__.py
@@ -1,2 +1,2 @@
 __author__ = "Sergei Y. Bogdanov <syubogdanov@outlook.com>"
-__version__ = "0.2.0-rc"
+__version__ = "0.2.0"