+++ CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Added `backlib.py310.json`;
+* Added `backlib.py310.operator`;
 * Added `backlib.py311.json`;
+* Added `backlib.py311.operator`;
 * Added `backlib.py312.json`;
+* Added `backlib.py312.operator`;
 * Added `backlib.py313.operator`.
 
 ## [0.1.0] - 2025-03-11