+++ CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Added `backlib.py310.builtins`;
 * Added `backlib.py310.errno`;
 * Added `backlib.py310.json`;
 * Added `backlib.py310.operator`;
 * Added `backlib.py310.os`;
 * Added `backlib.py310.os.path`;
 * Added `backlib.py310.stat`;
+* Added `backlib.py311.builtins`;
 * Added `backlib.py311.errno`;
 * Added `backlib.py311.json`;
 * Added `backlib.py311.operator`;
@@ -20,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `backlib.py311.os.path`;
 * Added `backlib.py311.stat`;
 * Added `backlib.py311.tomllib`;
+* Added `backlib.py312.builtins`;
 * Added `backlib.py312.errno`;
 * Added `backlib.py312.json`;
 * Added `backlib.py312.operator`;
@@ -27,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `backlib.py312.os.path`;
 * Added `backlib.py312.stat`;
 * Added `backlib.py312.tomllib`;
+* Added `backlib.py313.builtins`;
 * Added `backlib.py313.operator`;
 * Added `backlib.py313.os`;
 * Added `backlib.py310.os.path`.