+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-03-19
+
 * Added `backlib.py310.builtins`;
 * Added `backlib.py310.errno`;
 * Added `backlib.py310.json`;